[Ticket #12538] Access to data from outside the firewall should be through Web Components
Changed the maxReceivedMessageSize configuration option for WCF services to be
a long instead of an integer.  This required in a few places doing Math.Min() to
make sure the setting has a max value of int.MaxValue.  There are several places
the MaxReceivedMessage size is used for settings that are limited to int.MaxValue
as the maximum value.

--HG--
branch : Polaris

+++ Enterprise/Common/ServiceChannelConfigurationArgs.cs
@@ -38,7 +38,7 @@ namespace ClearCanvas.Enterprise.Common
             Type channelFactoryClass,
             Uri serviceUri,
             bool authenticationRequired,
-            int maxReceivedMessageSize,
+            long maxReceivedMessageSize,
             X509CertificateValidationMode certificateValidationMode,
             X509RevocationMode revocationMode)
         {
@@ -70,7 +70,7 @@ namespace ClearCanvas.Enterprise.Common
 		/// <summary>
 		/// The maximum size of received messages to allow, in bytes.
 		/// </summary>
-        public int MaxReceivedMessageSize;
+        public long MaxReceivedMessageSize;
 
 		/// <summary>
 		/// The time, in seconds, in which a send operation must complete.
